relocate carousel and return to top with other directory components
squash
